ipa-kdb: avoid undefined behavior when formatting SID identifier authority
dom_sid_string() shifted a uint8_t (promoted to signed int) left by
24 bits without first widening it to an unsigned type; if the byte's
high bit is set, the shift overflows a signed int, which is undefined
behavior in C. Cast each byte to uint32_t before shifting, matching
the identifier-authority reconstruction already used elsewhere.

Related: https://codeberg.org/freeipa/freeipa/issues/10033

+++ daemons/ipa-kdb/ipa_kdb_mspac.c
@@ -190,10 +190,10 @@ char *dom_sid_string(TALLOC_CTX *memctx, const struct dom_sid *dom_sid)
         return NULL;
     }
 
-    ia = (dom_sid->id_auth[5]) +
-         (dom_sid->id_auth[4] << 8 ) +
-         (dom_sid->id_auth[3] << 16) +
-         (dom_sid->id_auth[2] << 24);
+    ia = ((uint32_t)dom_sid->id_auth[5]) +
+         ((uint32_t)dom_sid->id_auth[4] << 8 ) +
+         ((uint32_t)dom_sid->id_auth[3] << 16) +
+         ((uint32_t)dom_sid->id_auth[2] << 24);
 
     ofs = snprintf(buf, len, "S-%u-%lu", (unsigned int) dom_sid->sid_rev_num,
                                             (unsigned long) ia);
